requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Database
sqlalchemy==2.0.23

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Visualization
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging
python-json-logger==2.0.7
